+++ lib/core/src/display/render/css3d/object.rs
@@ -1,6 +1,6 @@
 use crate::prelude::*;
 
-use crate::display::rendering::Transform;
+use crate::display::render::css3d::Transform;
 
 use nalgebra::UnitQuaternion;
 use nalgebra::Vector3;
