+++ lib/core/src/display/world/scene.rs
