+++ lib/core/src/display/world/scene/mesh_registry.rs
