+++ lib/core/tests/html_renderer.rs
@@ -18,10 +18,10 @@ extern "C" {
 
 #[cfg(test)]
 mod tests {
-    use basegl::display::rendering::Scene;
-    use basegl::display::rendering::Camera;
-    use basegl::display::rendering::html::HTMLObject;
-    use basegl::display::rendering::html::HTMLRenderer;
+    use basegl::display::css3d::Scene;
+    use basegl::display::css3d::Camera;
+    use basegl::display::css3d::html::HTMLObject;
+    use basegl::display::css3d::html::HTMLRenderer;
     use basegl::system::web::StyleSetter;
     use basegl::system::web::get_performance;
     use web_test::*;
